move out all helper functions
